use Ducks pettern
